requirements.txt
flask
flask_wtf
wtforms
wtforms.validators
requests
requests_cache
cassandra-driver
